Avoiding more compilation warnings

+++ src/io/formats/sogs.cpp
@@ -430,7 +430,7 @@ namespace lfs::io {
                                                                                                                        : 0);
 
                     int num_coeffs = SH_COEFFS[sh_degree];
-                    int palette_size = sh_meta.palette_size > 0 ? sh_meta.palette_size : centroids_img.size() / (64 * num_coeffs * 4);
+                    int palette_size = static_cast<int>(sh_meta.palette_size > 0 ? sh_meta.palette_size : centroids_img.size() / (64 * num_coeffs * 4));
 
                     LOG_DEBUG("Decoding SH: degree={}, coeffs={}, palette_size={}",
                               sh_degree, num_coeffs, palette_size);
@@ -1071,7 +1071,7 @@ namespace lfs::io {
             quats[ti * 4 + 0] = static_cast<uint8_t>(255.0f * (q[other_idx[0]] * 0.5f + 0.5f));
             quats[ti * 4 + 1] = static_cast<uint8_t>(255.0f * (q[other_idx[1]] * 0.5f + 0.5f));
             quats[ti * 4 + 2] = static_cast<uint8_t>(255.0f * (q[other_idx[2]] * 0.5f + 0.5f));
-            quats[ti * 4 + 3] = 252 + max_comp;
+            quats[ti * 4 + 3] = static_cast<uint8_t>(252 + max_comp);
         }
 
         if (auto result = write_webp("quats.webp", quats.data(), width, height); !result) {
